Moved NumberPresentation enum from ICU4N.Text to ICU4N.Globalization. Changed the values to start with 0 instead of 1 so we have a reasonable default value in .NET.
