+++ src/ICU4N/Text/RuleBasedNumberFormat.cs
@@ -22,90 +22,6 @@ using StringBuffer = System.Text.StringBuilder;
 namespace ICU4N.Text
 {
     /// <summary>
-    /// The presentation of <see cref="RuleBasedNumberFormat"/>.
-    /// </summary>
-#if FEATURE_LEGACY_NUMBER_FORMAT
-    public
-#else
-    internal
-#endif
-        enum NumberPresentation
-    {
-        /// <summary>
-        /// Indicates to create a spellout formatter that spells out a value
-        /// in words in the desired language.
-        /// </summary>
-        /// <draft>ICU 60.1</draft>
-        SpellOut = 1,
-
-        /// <summary>
-        /// Indicates to create an ordinal formatter that attaches an ordinal
-        /// suffix from the desired language to the end of the number (e.g. "123rd").
-        /// </summary>
-        /// <draft>ICU 60.1</draft>
-        Ordinal = 2,
-
-        /// <summary>
-        /// Indicates to create a duration formatter that formats a duration in
-        /// seconds as hours, minutes, and seconds.
-        /// </summary>
-        /// <draft>ICU 60.1</draft>
-        Duration = 3,
-
-        /// <summary>
-        /// Indicates to create a numbering system formatter to format a number in
-        /// a rules-based numbering system such as <c>%hebrew</c> for Hebrew numbers or <c>%roman-upper</c>
-        /// for upper-case Roman numerals.
-        /// </summary>
-        /// <draft>ICU 60.1</draft>
-        NumberingSystem = 4,
-    }
-
-    /// <summary>
-    /// Extensions to <see cref="NumberPresentation"/>.
-    /// </summary>
-    internal static class NumberPresentationExtensions
-    {
-        internal static bool IsDefined(this NumberPresentation presentation)
-            => presentation >= NumberPresentation.SpellOut && presentation <= NumberPresentation.NumberingSystem;
-
-        /// <summary>
-        /// Gets the rule name key to lookup this rule in the ICU resources.
-        /// <para/>
-        /// This is used internally to lookup the resource data.
-        /// </summary>
-        /// <param name="presentation">This <see cref="NumberPresentation"/> value.</param>
-        /// <returns>The rule name key to lookup this rule in the ICU resources.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">The <paramref name="presentation"/> value was not recognized.</exception>
-        internal static string ToRuleNameKey(this NumberPresentation presentation) => presentation switch // ICU4N: This was originally in RuleBasedNumberFormat
-        {
-            NumberPresentation.SpellOut => "RBNFRules/SpelloutRules",
-            NumberPresentation.Ordinal => "RBNFRules/OrdinalRules",
-            NumberPresentation.Duration => "RBNFRules/DurationRules",
-            NumberPresentation.NumberingSystem => "RBNFRules/NumberingSystemRules",
-            _ => throw new ArgumentOutOfRangeException(nameof(presentation), $"Not expected presentation value: {presentation}"),
-        };
-
-        /// <summary>
-        /// Gets the rule localizations key to lookup this rule in the ICU resources.
-        /// <para/>
-        /// This is used internally to lookup the resource data.
-        /// </summary>
-        /// <param name="presentation">This <see cref="NumberPresentation"/> value.</param>
-        /// <returns>The rule localizations key to lookup this rule in the ICU resources.</returns>
-        /// <exception cref="ArgumentOutOfRangeException">The <paramref name="presentation"/> value was not recognized.</exception>
-        internal static string ToRuleLocalizationsKey(this NumberPresentation presentation) => presentation switch // ICU4N: This was originally in RuleBasedNumberFormat
-        {
-            NumberPresentation.SpellOut => "SpelloutLocalizations",
-            NumberPresentation.Ordinal => "OrdinalLocalizations",
-            NumberPresentation.Duration => "DurationLocalizations",
-            NumberPresentation.NumberingSystem => "NumberingSystemLocalizations",
-            _ => throw new ArgumentOutOfRangeException(nameof(presentation), $"Not expected presentation value: {presentation}"),
-        };
-    }
-
-
-    /// <summary>
     /// A class that formats numbers according to a set of rules. This number formatter is
     /// typically used for spelling out numeric values in words (e.g., 25,3476 as
     /// "twenty-five thousand three hundred seventy-six" or "vingt-cinq mille trois
@@ -883,7 +799,7 @@ namespace ICU4N.Text
 
             try
             {
-                ICUResourceBundle rules = bundle.GetWithFallback("RBNFRules/" + rulenames[(int)format - 1]);
+                ICUResourceBundle rules = bundle.GetWithFallback(format.ToRuleNameKey());
                 UResourceBundleEnumerator it = rules.GetEnumerator();
                 while (it.MoveNext())
                 {
@@ -896,7 +812,7 @@ namespace ICU4N.Text
 
             // We use findTopLevel() instead of get() because
             // it's faster when we know that it's usually going to fail.
-            UResourceBundle locNamesBundle = bundle.FindTopLevel(locnames[(int)format - 1]);
+            UResourceBundle locNamesBundle = bundle.FindTopLevel(format.ToRuleLocalizationsKey());
             if (locNamesBundle != null)
             {
                 localizations = new string[locNamesBundle.Length][];
@@ -909,13 +825,6 @@ namespace ICU4N.Text
 
             Init(description.ToString(), localizations);
         }
-
-        private static readonly string[] rulenames = {
-            "SpelloutRules", "OrdinalRules", "DurationRules", "NumberingSystemRules",
-        };
-        private static readonly string[] locnames = {
-            "SpelloutLocalizations", "OrdinalLocalizations", "DurationLocalizations", "NumberingSystemLocalizations",
-        };
 
         /// <summary>
         /// Creates a <see cref="RuleBasedNumberFormat"/> from a predefined description. Uses the
