+++ tests/ICU4N.Tests/Dev/Test/Format/RBNFParseTest.cs
@@ -1,4 +1,5 @@
-﻿using ICU4N.Text;
+﻿using ICU4N.Globalization;
+using ICU4N.Text;
 using J2N.Numerics;
 using NUnit.Framework;
 using System;
